Move RhoCmds to their own files
This only affects the Python module structure of the program, not the
Ansible modules. However, this change is a prerequisite for changing
the Ansible module structure, if we want to do that, and also makes
the code cleaner and easier to read.

I've tried to keep this PR completely mechanical - just move code from
run_cmds.py to other files and update names and include statements
appropriately.